feat: kafka, outbox

+++ point-service/src/main/java/com/oneonone/pointservice/application/command/CreatePointCommand.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class CreatePointCommand {
     private final PointType type;
-    private final int amount;
+    private final Long amount;
     private final String description;
     private final Long userId;
 }
